# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/drone/": {
            "get": {
                "description": "Get all registered drones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Get Drones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.DroneRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new drone entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Create a new drone",
                "parameters": [
                    {
                        "description": "Request of Creating Drone",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DroneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.DroneRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/drone/load": {
            "post": {
                "description": "Loading a drone with medication items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Load Medications",
                "parameters": [
                    {
                        "description": "Request of load medications",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LoadMedicationsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.DroneRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/drone/{serial_number}/medications": {
            "get": {
                "description": "checking loaded medication items for a given drone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drone"
                ],
                "summary": "Get Loaded Medications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Serial Number",
                        "name": "serial_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.MedicationRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.DroneModelEnum": {
            "type": "string",
            "enum": [
                "Light",
                "Middle",
                "Cruiser",
                "Heavy"
            ],
            "x-enum-varnames": [
                "Light",
                "Middle",
                "Cruiser",
                "Heavy"
            ]
        },
        "entity.DroneRequest": {
            "type": "object",
            "properties": {
                "battery_capacity": {
                    "type": "integer"
                },
                "drone_model": {
                    "$ref": "#/definitions/entity.DroneModelEnum"
                },
                "serial_number": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/entity.DroneStateEnum"
                },
                "weight_limit": {
                    "type": "number"
                }
            }
        },
        "entity.DroneStateEnum": {
            "type": "string",
            "enum": [
                "IDLE",
                "LOADING",
                "LOADED",
                "DELIVERING",
                "DELIVERED",
                "RETURNING"
            ],
            "x-enum-varnames": [
                "IDLE",
                "LOADING",
                "LOADED",
                "DELIVERING",
                "DELIVERED",
                "RETURNING"
            ]
        },
        "entity.LoadMedicationsRequest": {
            "type": "object",
            "properties": {
                "medications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MedicationRequest"
                    }
                },
                "serial_number": {
                    "type": "string"
                }
            }
        },
        "entity.MedicationRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        }
    }
}